+++ src/main/java/com/jimin/readingjournal/domain/auth/controller/MainController.java
@@ -1,4 +1,4 @@
-package com.jimin.readingjournal.controller;
+package com.jimin.readingjournal.domain.auth.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
