--- src/main/java/com/jimin/readingjournal/mapper/ReadingJournalMapper.java
@@ -1,11 +1,0 @@
-package com.jimin.readingjournal.mapper;
-
-import com.jimin.readingjournal.dto.UserDto;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface ReadingJournalMapper {
-    void insertUser(UserDto userDto);
-
-    UserDto selectUserByUserId(String id);
-}
