--- src/main/java/com/jimin/readingjournal/service/AuthService.java
@@ -1,11 +1,0 @@
-package com.jimin.readingjournal.service;
-
-import com.jimin.readingjournal.request.SigninReq;
-import com.jimin.readingjournal.request.SignupReq;
-import jakarta.servlet.http.HttpSession;
-
-public interface AuthService {
-    void signup(SignupReq signupReq);
-
-    boolean signin(SigninReq signinReq, HttpSession session);
-}
