+++ src/main/resources/mapper/auth-mapper.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.jimin.readingjournal.mapper.ReadingJournalMapper">
-    <insert id="insertUser" parameterType="com.jimin.readingjournal.dto.UserDto">
+<mapper namespace="com.jimin.readingjournal.domain.auth.mapper.AuthMapper">
+    <insert id="insertUser" parameterType="com.jimin.readingjournal.domain.auth.dto.UserDto">
         INSERT INTO users (user_id, user_pw)
         VALUES (#{userId}, #{password})
     </insert>
@@ -11,7 +11,7 @@
         SELECT * FROM users
         WHERE user_id=#{userId};
     </select>
-    <select id="selectUserByUserIdAndUserPassword" parameterType="com.jimin.readingjournal.dto.UserDto">
+    <select id="selectUserByUserIdAndUserPassword" parameterType="com.jimin.readingjournal.domain.auth.dto.UserDto">
         SELECT * FROM users
         WHERE user_id=#{userId} and user_pw=#{password};
     </select>
